+++ src/QueryType/ManagedResources/ResponseParser/Stopword.php
@@ -44,8 +44,6 @@ class Stopword extends ResponseParserAbstract implements ResponseParserInterface
             $parsed['items'] = $items;
         }
 
-        $parsed = $this->addHeaderInfo($data, $parsed);
-
         return $parsed;
     }
 }