+++ src/QueryType/ManagedResources/ResponseParser/Stopwords.php
@@ -52,8 +52,6 @@ class Stopwords extends ResponseParserAbstract implements ResponseParserInterfac
             }
         }
 
-        $parsed = $this->addHeaderInfo($data, $parsed);
-
         return $parsed;
     }
 }