+++ src/QueryType/ManagedResources/ResponseParser/Synonym.php
@@ -47,8 +47,6 @@ class Synonym extends ResponseParserAbstract implements ResponseParserInterface
             $parsed['items'] = $items;
         }
 
-        $parsed = $this->addHeaderInfo($data, $parsed);
-
         return $parsed;
     }
 }