+++ src/QueryType/Spellcheck/ResponseParser.php
@@ -12,7 +12,6 @@ namespace Solarium\QueryType\Spellcheck;
 use Solarium\Core\Query\AbstractResponseParser as ResponseParserAbstract;
 use Solarium\Core\Query\ResponseParserInterface;
 use Solarium\Core\Query\Result\ResultInterface;
-use Solarium\QueryType\Spellcheck\Result\Result;
 
 /**
  * Parse Spellcheck response data.
@@ -62,14 +61,11 @@ class ResponseParser extends ResponseParserAbstract implements ResponseParserInt
             }
         }
 
-        return $this->addHeaderInfo(
-            $data,
-            [
-                'results' => $suggestions,
-                'all' => $allSuggestions,
-                'collation' => $collation,
-            ]
-        );
+        return [
+            'results' => $suggestions,
+            'all' => $allSuggestions,
+            'collation' => $collation,
+        ];
     }
 
     /**