+++ src/QueryType/Suggester/ResponseParser.php
@@ -48,13 +48,10 @@ class ResponseParser extends ResponseParserAbstract implements ResponseParserInt
             }
         }
 
-        return $this->addHeaderInfo(
-            $data,
-            [
-                'results' => $dictionaries,
-                'all' => $allSuggestions,
-            ]
-        );
+        return [
+            'results' => $dictionaries,
+            'all' => $allSuggestions,
+        ];
     }
 
     /**