+++ tests/QueryType/Extract/ResultTest.php
@@ -19,7 +19,6 @@ class ExtractResultDummy extends ExtractResult
 
     public function __construct()
     {
-        $this->status = 1;
-        $this->queryTime = 12;
+        $this->responseHeader = ['status' => 1, 'QTime' => 12];
     }
 }