+++ tests/QueryType/Ping/ResponseParserTest.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Solarium\Tests\QueryType\Update;
+namespace Solarium\Tests\QueryType\Ping;
 
 use PHPUnit\Framework\TestCase;
 use Solarium\Core\Client\Response;
-use Solarium\QueryType\Update\Query\Query;
-use Solarium\QueryType\Update\ResponseParser;
-use Solarium\QueryType\Update\Result;
+use Solarium\QueryType\Ping\Query;
+use Solarium\QueryType\Ping\ResponseParser;
+use Solarium\QueryType\Ping\Result;
 
 class ResponseParserTest extends TestCase
 {
     public function testParse()
     {
-        $data = '{}';
+        $data = '{"status":"OK"}';
 
         $response = new Response($data, ['HTTP 1.1 200 OK']);
         $result = new Result(new Query(), $response);
         $parser = new ResponseParser();
         $parsed = $parser->parse($result);
 
-        $this->assertSame([], $parsed);
+        $this->assertSame('OK', $parsed['status']);
     }
 }